31/01

+++ serverapp/Session/JWT.cs
@@ -5,6 +5,6 @@
         public string Key { get; set; }
         public string Issuer { get; set; }
         public string Audience { get; set; }
-        public int Expiration { get; set; }
+        public double Expiration { get; set; }
     }
 }
